Create Age_Calculator
Age Calculator

Project Overview

The Age Calculator is a simple yet useful application that allows users to calculate their age based on their date of birth. This project is a great way to practice and demonstrate skills in HTML, CSS, and JavaScript. The application takes the user's date of birth as input and calculates the exact age in years, months, and days.

Features

1.User-friendly interface with input validation.
2.Accurate calculation of age in years, months, and days.
3.Responsive design that works on various devices.

Technologies Used

1.HTML5: For structuring the application.
2.CSS3: For styling the application and making it visually appealing.
3.JavaScript: For implementing the core functionality of age calculation